requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
exceptiongroup==1.1.1
frozenlist==1.3.0
genshin @ git+https://github.com/thesadru/genshin.py.git@202dc9ea2580623aa493eb6a051c7b2ec560a95b # TODO replace with the commit with pure fiction hotfix
greenlet==2.0.2
h11==0.13.0
httpcore==0.17.2
httpx==0.24.1
idna==3.3
multidict==6.0.2
nextcord==2.4.2
outcome==1.2.0
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.13
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
wsproto==1.1.0
yarl==1.7.2
